# Optimize HIR type data structure