--- a/lib/hir/src/types/map.rs
+++ b/lib/hir/src/types/map.rs
@@ -4,35 +4,47 @@ use serde::{Deserialize, Serialize};
 use std::sync::Arc;
 
 #[derive(Clone, Debug, Deserialize, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize)]
-pub struct Map {
-    key: Arc<Type>,
-    value: Arc<Type>,
+pub struct Map(Arc<MapInner>);
+
+#[derive(Clone, Debug, Deserialize, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize)]
+struct MapInner {
+    key: Type,
+    value: Type,
     position: Position,
 }
 
 impl Map {
     pub fn new(key: impl Into<Type>, value: impl Into<Type>, position: Position) -> Self {
-        Self {
-            key: Arc::new(key.into()),
-            value: Arc::new(value.into()),
-            position,
-        }
+        Self(
+            MapInner {
+                key: (key.into()),
+                value: (value.into()),
+                position,
+            }
+            .into(),
+        )
     }
 
     pub fn key(&self) -> &Type {
-        &self.key
+        &self.0.key
     }
 
     pub fn value(&self) -> &Type {
-        &self.value
+        &self.0.value
     }
 
     pub fn position(&self) -> &Position {
-        &self.position
+        &self.0.position
     }
 
-    pub fn set_position(mut self, position: Position) -> Self {
-        self.position = position;
-        self
+    pub fn set_position(&self, position: Position) -> Self {
+        Self(
+            MapInner {
+                key: (self.0.key.clone()),
+                value: (self.0.value.clone()),
+                position,
+            }
+            .into(),
+        )
     }
 }