--- a/lib/hir/src/types/record.rs
+++ b/lib/hir/src/types/record.rs
@@ -1,8 +1,12 @@
 use position::Position;
 use serde::{Deserialize, Serialize};
+use std::sync::Arc;
 
 #[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq, PartialOrd, Ord, Serialize)]
-pub struct Record {
+pub struct Record(Arc<RecordInner>);
+
+#[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq, PartialOrd, Ord, Serialize)]
+struct RecordInner {
     name: String,
     original_name: String,
     position: Position,
@@ -14,27 +18,36 @@ impl Record {
         original_name: impl Into<String>,
         position: Position,
     ) -> Self {
-        Self {
-            name: name.into(),
-            original_name: original_name.into(),
-            position,
-        }
+        Self(
+            RecordInner {
+                name: name.into().into(),
+                original_name: original_name.into().into(),
+                position,
+            }
+            .into(),
+        )
     }
 
     pub fn name(&self) -> &str {
-        &self.name
+        &self.0.name
     }
 
     pub fn original_name(&self) -> &str {
-        &self.original_name
+        &self.0.original_name
     }
 
     pub fn position(&self) -> &Position {
-        &self.position
+        &self.0.position
     }
 
-    pub fn set_position(mut self, position: Position) -> Self {
-        self.position = position;
-        self
+    pub fn set_position(&self, position: Position) -> Self {
+        Self(
+            RecordInner {
+                name: self.0.name.clone(),
+                original_name: self.0.original_name.clone(),
+                position,
+            }
+            .into(),
+        )
     }
 }