… lib/hir-mir/src/transformation/equal_operation/snapshots/hir_mir__transformation__equal_operation__module__tests__transform_comparable_type.snap
… lib/hir-mir/src/transformation/hash_calculation/snapshots/hir_mir__transformation__hash_calculation__module__tests__transform_comparable_type.snap
… lib/hir-mir/src/transformation/map_context/snapshots/hir_mir__transformation__map_context__module__tests__transform_function_value.snap
… lib/hir-mir/src/transformation/map_context/snapshots/hir_mir__transformation__map_context__module__tests__transform_none_key_and_none_value.snap
… lib/hir-mir/src/transformation/snapshots/hir_mir__transformation__if_list__tests__transform_if_list_with_any_type.snap
… lib/hir-mir/src/transformation/snapshots/hir_mir__transformation__if_list__tests__transform_if_list_with_number_type.snap
… lib/hir-mir/src/transformation/snapshots/hir_mir__transformation__if_map__tests__transform_if_map.snap
… lib/hir-mir/src/transformation/snapshots/hir_mir__transformation__map_literal__tests__transform_empty_map.snap
… lib/hir-mir/src/transformation/snapshots/hir_mir__transformation__map_literal__tests__transform_empty_map_with_function_value.snap
… lib/hir-mir/src/transformation/snapshots/hir_mir__transformation__map_literal__tests__transform_map_with_2_entries.snap
… lib/hir-mir/src/transformation/snapshots/hir_mir__transformation__map_literal__tests__transform_map_with_entry.snap
… lib/hir-mir/src/transformation/snapshots/hir_mir__transformation__map_literal__tests__transform_map_with_spread_map.snap
… lib/parse/src/snapshots/parse__parsers__tests__expressions__call__fail_to_parse_call.snap
… lib/parse/src/snapshots/parse__parsers__tests__import__fail_to_parse_private_external_module_directory.snap
… lib/parse/src/snapshots/parse__parsers__tests__import__fail_to_parse_private_external_module_file.snap